// gds-frontend/src/gs-api/model/articleDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CategoryDto } from './categoryDto';


export interface ArticleDto { 
    id?: number;
    codeArticle?: string;
    designation?: string;
    prixUnitaireHt?: number;
    tauxTva?: number;
    prixUnitaireTtc?: number;
    photo?: string;
    category?: CategoryDto;
    entrepriseId?: number;
}