// gds-frontend/src/gs-api/model/clientDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdresseDto } from './adresseDto';


export interface ClientDto { 
    id?: number;
    nom?: string;
    prenom?: string;
    adresse?: AdresseDto;
    photo?: string;
    email?: string;
    numTel?: string;
    entrepriseId?: number;
}