// gds-frontend/src/gs-api/model/commandeClientDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ClientDto } from './clientDto';
import { LigneCommandeClientDto } from './ligneCommandeClientDto';


export interface CommandeClientDto { 
    id?: number;
    code?: string;
    dateCommande?: string;
    client?: ClientDto;
    ligneCommandeClient?: Array<LigneCommandeClientDto>;
    etatCommande?: CommandeClientDto.EtatCommandeEnum;
    entrepriseId?: number;
    commandeLivree?: boolean;
}
export namespace CommandeClientDto {
    export const EtatCommandeEnum = {
        EnPreparation: 'EN_PREPARATION',
        Validee: 'VALIDEE',
        Livree: 'LIVREE'
    } as const;
    export type EtatCommandeEnum = typeof EtatCommandeEnum[keyof typeof EtatCommandeEnum];
}