// gds-frontend/src/gs-api/model/commandeFournisseurDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LigneCommandeFournisseurDto } from './ligneCommandeFournisseurDto';
import { FournisseurDto } from './fournisseurDto';


export interface CommandeFournisseurDto { 
    id?: number;
    code?: string;
    dateCommande?: string;
    fournisseur?: FournisseurDto;
    ligneCommandeFournisseurs?: Array<LigneCommandeFournisseurDto>;
    etatCommande?: CommandeFournisseurDto.EtatCommandeEnum;
    entrepriseId?: number;
    commandeLivree?: boolean;
}
export namespace CommandeFournisseurDto {
    export const EtatCommandeEnum = {
        EnPreparation: 'EN_PREPARATION',
        Validee: 'VALIDEE',
        Livree: 'LIVREE'
    } as const;
    export type EtatCommandeEnum = typeof EtatCommandeEnum[keyof typeof EtatCommandeEnum];
}