// gds-frontend/src/gs-api/model/ligneCommandeClientDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CommandeClientDto } from './commandeClientDto';
import { ArticleDto } from './articleDto';


export interface LigneCommandeClientDto { 
    id?: number;
    article?: ArticleDto;
    commandeClient?: CommandeClientDto;
    quantite?: number;
    prixUnitaire?: number;
    entrepriseId?: number;
}